requirements.txt
openai>=1.0.0
PyPDF2>=3.0.1
Pillow
python-dotenv
requests